+++ client/src/views/processors/ProcessorsData.js
@@ -118,14 +118,16 @@ const ProcessorsData = () => {
           // })
           // setTableData(proArr)
           setTableData(response.data.result)
-          setIsloader(false)
         }
       }
+      setIsloader(false)
     }).catch(err => {
       toast.error(err.message, toast_options);
       // console.log("Record Issue Error Processor Data", err)
+      if(err.message.includes("403")){
       localStorage.clear();
       history.push("/");
+      }
     });
   }
   const handleClickToOpen = () => {
@@ -173,16 +175,18 @@ const ProcessorsData = () => {
           toast.error(response.data.err, toast_options);
         } else {
           // this.setState({ pdfImage: response.data.result, isLoading: false })
-          setIsloader(false)
           setTableData([])
           getProcessorData()
           handleToClose()
         }
+        setIsloader(false)
       }).catch(err => {
         toast.error(err.message, toast_options);
         console.log("Company Data Issue Error", err);
+        if(err.message.includes("403")){
         localStorage.clear();
         history.push("/");
+        }
       });
     }}
